+++ CardPayments/src/main/java/com/paypal/android/cardpayments/CardVaultResult.kt
@@ -1,5 +1,7 @@
 package com.paypal.android.cardpayments
 
+import androidx.annotation.RestrictTo
+
 /**
  * A result returned by [CardClient] when an a successful vault occurs.
  *
@@ -8,7 +10,7 @@ package com.paypal.android.cardpayments
  * @property [didAttemptThreeDSecureAuthentication] 3DS verification was attempted.
  * Use v2/checkout/orders/{orderId} in your server to get verification results.
  */
-data class CardVaultResult internal constructor(
+data class CardVaultResult @RestrictTo(RestrictTo.Scope.LIBRARY_GROUP) constructor(
     val setupTokenId: String,
     val status: String? = null,
     val didAttemptThreeDSecureAuthentication: Boolean = false